(update) upgrade to react-router 4
