+++ package.json
@@ -53,17 +53,18 @@
     "webpack-dev-server": "^2.2.1"
   },
   "dependencies": {
-    "babel-polyfill": "^6.16.0",
-    "es6-promise": "^3.1.2",
+    "babel-polyfill": "^6.23.0",
+    "es6-promise": "^3.3.1",
     "immutable": "^3.8.1",
     "isomorphic-fetch": "^2.2.1",
-    "prop-types": "^15.5.8",
-    "react": "^15.5.0",
-    "react-dom": "^15.5.0",
-    "react-redux": "^4.4.0",
-    "react-router": "^3.0.0",
-    "redux": "^3.3.1",
-    "redux-thunk": "^2.0.1"
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^4.4.8",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1",
+    "redux": "^3.6.0",
+    "redux-thunk": "^2.2.0"
   },
   "description": "Starter boilerplate for React and Redux, using Webpack 2",
   "repository": {
